vsezved: Get data on Czech schools from <https://stistko.uiv.cz/registr/> and <https://data.msmt.cz/>

Get access to data on Czech schools: the register open data provided by the Ministry of Education at https://data.msmt.cz/ and non-open web database at http://stistko.uiv.cz/registr/ This is mostly organisatoinal data on primary and secondary schools.

See also

Author

Maintainer: Petr Bouchal pbouchal@gmail.com (ORCID)