Get Stistko Ciselnik (df from code)

Reads and processes the HTML file of a Stistko ciselnik based on a code

Usage

vz_get_codelist(code, dest_dir = NULL)

Arguments

code

A character string representing the code of the codelist.

Value

A data frame containing the processed data from the ciselnik.

Examples

vz_get_codelist("BASO")
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 58 Columns: 6
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (5): kod, txt, zkr, datzp, datkp
#> lgl (1): spc
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 57 × 6
#>    kod   txt                 zkr   spc   datzp               datkp 
#>    <chr> <chr>               <chr> <chr> <dttm>              <dttm>
#>  1 11    Matematické obory   Mate… NA    1998-09-01 00:00:00 NA    
#>  2 12    Geologické obory    Geol… NA    1998-09-01 00:00:00 NA    
#>  3 13    Geografické obory   Geog… NA    1998-09-01 00:00:00 NA    
#>  4 14    Chemické obory      Chem… NA    1998-09-01 00:00:00 NA    
#>  5 15    Biologické obory    Biol… NA    1998-09-01 00:00:00 NA    
#>  6 16    Ekologie a ochrana… Ekol… NA    1998-09-01 00:00:00 NA    
#>  7 17    Fyzikální obory     Fyzi… NA    1998-09-01 00:00:00 NA    
#>  8 18    Informatické obory  Info… NA    1999-09-01 00:00:00 NA    
#>  9 21    Hornictví a hornic… Horn… NA    1998-09-01 00:00:00 NA    
#> 10 23    Strojírenství a st… Stro… NA    1998-09-01 00:00:00 NA    
#> # ℹ 47 more rows